// frontend/src/api/models/handlerBalanceTransactionResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Balance transaction response
 */
export interface HandlerBalanceTransactionResponse {
  amount?: number
  balance_after?: number
  balance_before?: number
  created_at?: string
  customer_id?: string
  id?: string
  operator_id?: string
  reference?: string
  remark?: string
  source_id?: string
  source_type?: string
  tenant_id?: string
  transaction_date?: string
  transaction_type?: string
}