// frontend/src/api/models/handlerBalanceSummaryResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Customer balance summary response
 */
export interface HandlerBalanceSummaryResponse {
  current_balance?: number
  customer_id?: string
  total_consume?: number
  total_recharge?: number
  total_refund?: number
}