// frontend/src/api/models/listBalanceTransactionsParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { ListBalanceTransactionsSourceType } from './listBalanceTransactionsSourceType'
import type { ListBalanceTransactionsTransactionType } from './listBalanceTransactionsTransactionType'

export type ListBalanceTransactionsParams = {
  /**
   * Transaction type
   */
  transaction_type?: ListBalanceTransactionsTransactionType
  /**
   * Source type
   */
  source_type?: ListBalanceTransactionsSourceType
  /**
   * Start date (YYYY-MM-DD)
   */
  date_from?: string
  /**
   * End date (YYYY-MM-DD)
   */
  date_to?: string
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
}