// frontend/src/api/models/listBalanceTransactionsTransactionType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListBalanceTransactionsTransactionType =
  (typeof ListBalanceTransactionsTransactionType)[keyof typeof ListBalanceTransactionsTransactionType]

export const ListBalanceTransactionsTransactionType = {
  RECHARGE: 'RECHARGE',
  CONSUME: 'CONSUME',
  REFUND: 'REFUND',
  ADJUSTMENT: 'ADJUSTMENT',
  EXPIRE: 'EXPIRE',
} as const