// frontend/src/api/models/listBalanceTransactionsSourceType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListBalanceTransactionsSourceType =
  (typeof ListBalanceTransactionsSourceType)[keyof typeof ListBalanceTransactionsSourceType]

export const ListBalanceTransactionsSourceType = {
  MANUAL: 'MANUAL',
  SALES_ORDER: 'SALES_ORDER',
  SALES_RETURN: 'SALES_RETURN',
  RECEIPT_VOUCHER: 'RECEIPT_VOUCHER',
  SYSTEM: 'SYSTEM',
} as const